requirements.txt
pandas
selenium
beautifulsoup4
google-cloud-storage
google-auth
langchain
langchain_openai
langchain_community
tiktoken>=0.6.0  # otherwise the text-embedding-3-large model does not get recognized
python-dotenv
wget
openai
faiss-cpu
uvicorn
fastapi
argparse
tqdm
backoff